Feat: Moves decorators to external library
